+++ CHANGELOG.md
@@ -21,6 +21,7 @@ See [Issue Backlog](../../issues) and [Roadmap](../../milestones).
 - *Changed* Session Turnover UI to use DialogV2. [#302](https://github.com/Jagusti/fvtt-wfrp4e-gmtoolkit/issues/302)
 - *Changed* Add XP to suppress duplicate system chat messages for experience awards.
 - *Changed* Dark Whispers UI to use DialogV2. [#305](https://github.com/Jagusti/fvtt-wfrp4e-gmtoolkit/issues/305)
+- *Changed* Advantage Momentum UI to use DialogV2. [#306](https://github.com/Jagusti/fvtt-wfrp4e-gmtoolkit/issues/306)
 
 ## [Version 8.0.1](https://github.com/Jagusti/fvtt-wfrp4e-gmtoolkit/releases/tag/v8.0.1)  (2024-09-21)
 - *Added* WFRP4e v8 compatibility for Check Conditions macro.